refactor: move pkg/util/gpu to internal
Removes buildcfg dependency on pkg directory
