+++ cmd/internal/cli/actions_linux.go
@@ -26,6 +26,7 @@ import (
 	"github.com/hpcng/singularity/internal/pkg/util/bin"
 	"github.com/hpcng/singularity/internal/pkg/util/env"
 	"github.com/hpcng/singularity/internal/pkg/util/fs"
+	"github.com/hpcng/singularity/internal/pkg/util/gpu"
 	"github.com/hpcng/singularity/internal/pkg/util/shell/interpreter"
 	"github.com/hpcng/singularity/internal/pkg/util/starter"
 	"github.com/hpcng/singularity/internal/pkg/util/user"
@@ -38,7 +39,6 @@ import (
 	"github.com/hpcng/singularity/pkg/util/capabilities"
 	"github.com/hpcng/singularity/pkg/util/cryptkey"
 	"github.com/hpcng/singularity/pkg/util/fs/proc"
-	"github.com/hpcng/singularity/pkg/util/gpu"
 	"github.com/hpcng/singularity/pkg/util/namespaces"
 	"github.com/hpcng/singularity/pkg/util/rlimit"
 	"github.com/hpcng/singularity/pkg/util/singularityconf"
