+++ internal/pkg/runtime/engine/singularity/container_linux.go
@@ -26,6 +26,7 @@ import (
 	"github.com/hpcng/singularity/internal/pkg/util/fs/layout/layer/underlay"
 	"github.com/hpcng/singularity/internal/pkg/util/fs/mount"
 	fsoverlay "github.com/hpcng/singularity/internal/pkg/util/fs/overlay"
+	"github.com/hpcng/singularity/internal/pkg/util/gpu"
 	"github.com/hpcng/singularity/internal/pkg/util/mainthread"
 	"github.com/hpcng/singularity/internal/pkg/util/priv"
 	"github.com/hpcng/singularity/internal/pkg/util/user"
@@ -35,7 +36,6 @@ import (
 	singularity "github.com/hpcng/singularity/pkg/runtime/engine/singularity/config"
 	"github.com/hpcng/singularity/pkg/sylog"
 	"github.com/hpcng/singularity/pkg/util/fs/proc"
-	"github.com/hpcng/singularity/pkg/util/gpu"
 	"github.com/hpcng/singularity/pkg/util/loop"
 	"github.com/hpcng/singularity/pkg/util/namespaces"
 	"github.com/hpcng/singularity/pkg/util/singularityconf"
