+++ internal/pkg/util/gpu/paths_test.go
